requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
pathlib==1.0.1
Pillow==9.5.0
Werkzeug==2.3.4
